Work on the new classes.
