--- src/main/java/Scanner.java
@@ -1,4 +1,0 @@
-
-public class Scanner {
-
-}
